move calls count tests to another class add test utils class
